Add Version Upgrade action handler
