+++ src/handlers/actions/index.ts
@@ -5,3 +5,4 @@ export { default as handleEditDomainAction } from './editDomain';
 export { default as handleTokenUnlockedAction } from './unlockToken';
 export { default as handleMoveFundsAction } from './moveFunds';
 export { default as handleEditColonyAction } from './editColony';
+export { default as handleVersionUpgradeAction } from './versionUpgrade';
